requirements.txt
# NF-e Destinadas - Dependências
# Instale com: pip install -r requirements.txt

# Criptografia e leitura do certificado .pfx
cryptography>=41.0.0

# HTTP com suporte a certificado mútuo (TLS)
requests>=2.31.0
urllib3>=2.0.0

# Assinatura XML (obrigatória para Manifestação do Destinatário)
lxml>=4.9.0
signxml>=3.2.0

# (Opcional) pyOpenSSL como fallback para leitura do .pfx
pyOpenSSL>=23.0.0
